requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
matplotlib
duckdb
lxml
beautifulsoup4
requests
google-generativeai
